requirement.txt
dearpygui
pyperclip
pyclip
pillow
g4f[all]
datetime
win32ui
win32gui
ctypes
uuid
threading
ftfy